@@ -1,5 +1,5 @@
 import React from 'react';
-import { Grid } from '@material-ui/core';
+import { Grid } from '@mui/material';
 import { ComposeModel } from '../../../Common/Models/Dashboard/ComposeModel';
 import { Compose } from './Compose';
 
