@@ -1,10 +1,10 @@
 import React from 'react';
-import { Redirect, Route } from 'react-router-dom';
+import { Navigate, Route } from 'react-router-dom';
 
 export const PublicRoute: React.FC<any> = ({ component, isAuthenticated, ...rest }: any) => (
   <Route {...rest} render={(props: any) => (
     !isAuthenticated
       ? React.createElement(component, props)
-      : <Redirect to={{ pathname: '/' }} />
+      : <Navigate to={{ pathname: '/' }} />
   )} />
 );
